+++ CHANGELOG.md
@@ -1,13 +1,17 @@
 ## 0.7.1
 
-* Update minimum Dart SDK to 1.23.0.
+* The `shelf_io` server now adds a `shelf.io.connection_info` field to
+  `Request.context`, which provides access to the underlying
+  `HttpConnectionInfo` object.
 
 ## 0.7.0
 
 * Give a return type to the `Handler` typedef. This may cause static warnings
   where there previously were none, but all handlers should have already been
   returning a `Response` or `Future<Response>`.
+
 * Remove `HijackCallback` and `OnHijackCallback` typedefs.
+
 * **Breaking**: Change type of `onHijack` in the `Request` constructor to take
   an argument of `StreamChannel`.